Extract structs out of app_config
