+++ packages/ciphernode/config/src/lib.rs
@@ -1,8 +1,13 @@
 mod app_config;
-mod generate_id;
+pub mod chain_config;
+pub mod contract;
+pub mod generate_id;
+pub mod rpc;
 mod store_keys;
 pub mod validation;
 mod yaml;
 
 pub use app_config::*;
+pub use contract::*;
+pub use rpc::*;
 pub use store_keys::*;
